requirements.txt
pypdf>=6.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.21
Pillow>=10.0.0
